# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Iroha API
  version: 0.4.1
  description: |
    Active-development contract for Iroha's private data cockpit.
    /api/v1 is still pre-1.0 and may evolve between minor releases. This
    document is the implementation target for release 0.4.1.
servers:
  - url: /
    description: Same-origin or reverse-proxy deployment
...
paths:
...
  /api/v1/briefing:
    get:
      tags: [system]
      parameters:
        - $ref: "#/components/parameters/Date"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Day briefing with versioned sections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Briefing"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/raw-files:
    get:
      tags: [imports]
      responses:
        "200":
          description: Most recent raw files.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RawFile"
    post:
      tags: [imports]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, source_kind]
              properties:
                file:
                  type: string
                  format: binary
                source_kind:
                  $ref: "#/components/schemas/SourceKind"
                uploaded_via:
                  $ref: "#/components/schemas/UploadSource"
      responses:
        "201":
          description: Raw file stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawFile"
        "200":
          description: File content was already stored; duplicate is true.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawFile"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/raw-files/{rawFileId}:
    get:
      tags: [imports]
      parameters:
        - $ref: "#/components/parameters/RawFileId"
      responses:
        "200":
          description: Raw file metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/imports:
    get:
      tags: [imports]
      responses:
        "200":
          description: Most recent import jobs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImportJob"
    post:
      tags: [imports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateImport"
      responses:
        "202":
          description: Import job queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/imports/{importId}:
    get:
      tags: [imports]
      parameters:
        - $ref: "#/components/parameters/ImportId"
      responses:
        "200":
          description: Import job status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/activities:
    get:
      tags: [activities]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: sport_type
          in: query
          schema: { type: string }
        - name: started_from
          in: query
          description: Inclusive lower bound for started_at.
          schema: { type: string, format: date-time }
        - name: started_to
          in: query
          description: Exclusive upper bound for started_at.
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/ScopeDate"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/Timezone"
        - name: min_distance_m
          in: query
          schema: { type: number }
        - name: max_distance_m
          in: query
          schema: { type: number }
      responses:
        "200":
          description: Activity page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityPage"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/activities/overview:
    get:
      tags: [activities]
      parameters:
        - name: recent
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 5 }
        - name: timezone
          in: query
          schema:
            { type: string, description: IANA timezone used for calendar days }
      responses:
        "200":
          description: Dashboard activity projection with canonical active days and recent rows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityOverview"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/activities/summary:
    get:
      tags: [activities]
      parameters:
        - name: year
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/ScopeDate"
        - $ref: "#/components/parameters/Scope"
        - name: sport
          in: query
          schema: { type: string }
        - name: timezone
          in: query
          schema:
            {
              type: string,
              description: IANA timezone used for calendar buckets,
            }
      responses:
        "200":
          description: Activity summary aggregates (totals, by year/month/sport).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivitySummary"
  /api/v1/activities/bounds:
    get:
      tags: [activities]
      parameters:
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Earliest and latest calendar date with a real activity record, capped at today.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DateBounds"
...
  /api/v1/activities/{activityId}:
    get:
      tags: [activities]
      parameters:
        - $ref: "#/components/parameters/ActivityId"
      responses:
        "200":
          description: Activity detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/activities/{activityId}/route:
    get:
      tags: [activities]
      parameters:
        - $ref: "#/components/parameters/ActivityId"
      responses:
        "200":
          description: Ordered route points.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoutePoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/activities/{activityId}/samplings:
    get:
      tags: [activities]
      parameters:
        - $ref: "#/components/parameters/ActivityId"
      responses:
        "200":
          description: Activity sampling records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sampling"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/activities/{activityId}/laps:
    get:
      tags: [activities]
      parameters:
        - $ref: "#/components/parameters/ActivityId"
      responses:
        "200":
          description: Activity laps.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lap"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/sleep:
    get:
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/ScopeDate"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/Timezone"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Sleep session page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepPage"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/sleep/overview:
    get:
      tags: [sleep]
      parameters:
        - name: recent
          in: query
          description: Number of most recent sleep rows used for the main-sleep averages.
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        "200":
          description: Server-owned sleep dashboard projection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepOverview"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/sleep/aggregates:
    get:
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/SleepGranularity"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/ScopeDate"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Sleep aggregate buckets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepAggregateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/sleep/bounds:
    get:
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Earliest and latest wake date with a real sleep session, capped at today.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DateBounds"
  /api/v1/sleep/{sleepId}:
    get:
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/SleepId"
      responses:
        "200":
          description: Sleep session detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sleep"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/sleep/{sleepId}/segments:
    get:
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/SleepId"
      responses:
        "200":
          description: Sleep stage segments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SleepSegment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/daily:
    get:
      tags: [daily]
      parameters:
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/ScopeDate"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/Timezone"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Daily summary page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPage"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/daily/dates:
    get:
      tags: [daily]
      parameters:
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Canonical dates represented by daily facts, activities, sleep wake dates, or dated media events in the requested timezone.
          content:
            application/json:
              schema:
                type: array
                items: { type: string, format: date }
  /api/v1/daily/bounds:
    get:
      tags: [daily]
      parameters:
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Earliest and latest calendar date represented by any domain, capped at today.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DateBounds"
  /api/v1/daily/aggregates:
    get:
      tags: [daily]
      parameters:
        - $ref: "#/components/parameters/Granularity"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/ScopeDate"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Daily aggregate buckets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyAggregateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/reports/monthly:
    get:
      tags: [system]
      parameters:
        - $ref: "#/components/parameters/ReportDate"
        - $ref: "#/components/parameters/Month"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Synchronous typed monthly report over all private canonical domains.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/reports/monthly-series:
    get:
      tags: [system]
      parameters:
        - $ref: "#/components/parameters/ReportDate"
        - $ref: "#/components/parameters/ReportEndMonth"
        - $ref: "#/components/parameters/ReportSeriesMonths"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Server-aggregated monthly reports with empty months omitted from the trend points.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReportSeries"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/expenses:
    get:
      tags: [expenses]
      parameters:
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/ScopeDate"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/Timezone"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: currency
          in: query
          schema:
            type: string
            enum: [JPY, USD, EUR, GBP]
        - name: category
          in: query
          schema:
            type: string
            enum:
              [
                food,
                groceries,
                transport,
                shopping,
                housing,
                utilities,
                health,
                entertainment,
                subscriptions,
                work,
                other,
              ]
      responses:
        "200":
          description: Active expense page ordered by occurred date descending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpensePage"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpense"
      responses:
        "201":
          description: Expense created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "200":
          description: Idempotent retry returned the existing expense.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /api/v1/expenses/bounds:
    get:
      tags: [expenses]
      parameters:
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Earliest and latest occurred-on date among active expenses, capped at today.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DateBounds"
  /api/v1/expenses/{expenseId}:
    get:
      tags: [expenses]
      parameters:
        - $ref: "#/components/parameters/ExpenseId"
      responses:
        "200":
          description: Expense record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
    put:
      tags: [expenses]
      parameters:
        - $ref: "#/components/parameters/ExpenseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceExpense"
      responses:
        "200":
          description: Expense replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
    delete:
      tags: [expenses]
      parameters:
        - $ref: "#/components/parameters/ExpenseId"
      responses:
        "204":
          description: Expense tombstone recorded; repeat deletes are also successful.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/media:
    get:
      tags: [media]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: status
          in: query
          schema: { type: string }
        - name: media_type
          in: query
          schema: { type: string }
        - name: family
          in: query
          schema: { type: string }
        - name: completed_year
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Media library page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  parameters:
    RawFileId:
      name: rawFileId
      in: path
      required: true
      schema: { type: string }
    ImportId:
      name: importId
      in: path
      required: true
      schema: { type: string }
...
    Date:
      name: date
      in: query
      required: true
      schema: { type: string, format: date }
...
    Timezone:
      name: timezone
      in: query
      schema: { type: string, default: Asia/Tokyo }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
    BadRequest:
      description: Request could not be parsed or validated.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    SourceKind:
      type: string
      enum: [apple_health_export, gpx, fit, tcx, strava_export]
    UploadSource:
      type: string
      enum: [web, telegram, cli, ios_bridge]
...
    Briefing:
      type: object
      required: [date, previous_date, next_date, sections]
      properties:
        date: { type: string, format: date }
        previous_date: { type: string, format: date }
        next_date: { type: string, format: date }
        sections:
          { type: array, items: { type: object, additionalProperties: true } }